[FLINK-24253][jdbc] Mark Dialect as Public and move converter out of internal package
