+++ flink-connectors/flink-connector-jdbc/src/test/java/org/apache/flink/connector/jdbc/converter/AbstractJdbcRowConverterTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.connector.jdbc.internal.converter;
+package org.apache.flink.connector.jdbc.converter;
 
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.types.logical.IntType;
